requirements.txt
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
firebase-admin==5.2.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.46.1
grpcio-status==1.46.1
httplib2==0.20.4
idna==3.3
msgpack==1.0.3
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
